feat: Pre-populate database with sample wordlist
Pre-populates the database on its initial creation with a sample wordlist called "Sample Wordlist". This list includes a few English to "Youth Slang" word pairs.

This change also introduces a `HomeScreenViewModel` to provide the list of wordlists to the `HomeScreen`. The UI is updated to reflect the data from the ViewModel, including selection logic for the lists.

Additional changes:
- Replaces the placeholder logo in wordlist items with an `ic_list_alt` icon.
- Removes the "Starred Pairs" checkbox from the `WordListsPanel` as it was redundant with an option in the `PracticePanel`.