chore: Remove theme from AndroidManifest
Removes the `android:theme="@style/Theme.Voclet"` attribute from the `activity` tag in `AndroidManifest.xml`. This is done because the splash screen theme is now handled by the `installSplashScreen` API.

Also, reorders the `installSplashScreen()` call in `MainActivity.kt` to be the first statement in `onCreate`, which is the recommended practice.